fix: circular dependency for client type
